requirements.txt
django~=4.1.13
djangorestframework~=3.15.1
mysqlclient
torch~=2.2.2
torchvision
facenet-pytorch
tensorflow~=2.17.0
cmake
dlib
face_recognition
matplotlib~=3.9.1

numpy~=1.26.4
pillow~=10.2.0
scikit-learn~=1.5.1
opencv-python~=4.10.0.84
imutils